finish chapter 12: Module Product

+++ src/main/java/vn/hoidanit/laptopshop/controller/admin/UserController.java
@@ -16,29 +16,40 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.SessionAttributes;
 import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import jakarta.validation.Valid;
+import vn.hoidanit.laptopshop.domain.Role;
 import vn.hoidanit.laptopshop.domain.User;
 import vn.hoidanit.laptopshop.service.FileService;
+import vn.hoidanit.laptopshop.service.RoleService;
 import vn.hoidanit.laptopshop.service.FileService.Type;
 import vn.hoidanit.laptopshop.service.UserService;
 
 @Controller
+@SessionAttributes("roles")
 public class UserController {
 
   final private UserService userService;
   final private FileService fileService;
+  final private RoleService roleService;
 
   @InitBinder
   public void initBinder(WebDataBinder binder) {
     binder.registerCustomEditor(String.class, new StringTrimmerEditor(true)); // Converts "" to null
   }
 
-  public UserController(UserService userService, FileService fileService) {
+  @ModelAttribute("roles")
+  public List<Role> populateRoles() {
+    return roleService.getAll();
+  }
+
+  public UserController(UserService userService, FileService fileService, RoleService roleService) {
     this.userService = userService;
     this.fileService = fileService;
+    this.roleService = roleService;
   }
 
   // @RequestMapping("/")
@@ -62,19 +73,23 @@ public class UserController {
   @RequestMapping("/admin/user/create")
   public String getCreateUserPage(Model model) {
     model.addAttribute("newUser", new User());
+    // Không cần truyền xuống mỗi lân request nx do đã lưu trong session bằng
+    // @SessionAttribute
+    // model.addAttribute("roles", roleService.getAll());
     return "admin/user/create";
   }
 
   @RequestMapping(value = "/admin/user/create", method = RequestMethod.POST)
   public String createUser(@Valid @ModelAttribute("newUser") User user, BindingResult userBindingResult,
-      @RequestParam("avatar_file") MultipartFile file) {
+      @RequestParam("avatar_file") MultipartFile file,
+      RedirectAttributes redirectAttributes) {
 
     if (userService.emailExists(user.getEmail()))
       userBindingResult.rejectValue("email", "Unique", "Email already exists");
 
     if (userBindingResult.hasFieldErrors()) {
       List<FieldError> errors = userBindingResult.getFieldErrors();
-      errors.forEach((e) -> System.out.println(">>> ERR: " + e.getField() + " - " + e.getDefaultMessage()));
+      errors.forEach((e) -> System.out.println(">>> ERR Create User: " + e.getField() + " - " + e.getDefaultMessage()));
       return "admin/user/create"; // return view -> dữ liệu input vẫn còn lưu
       // return "redirect:/admin/user/create"; // redirect thì như refresh lại vào lại
       // url đó -> dữ liệu input bị mất
@@ -85,6 +100,7 @@ public class UserController {
       user.setAvatar(savedFile);
     }
     User newUser = userService.registerUser(user);
+    redirectAttributes.addFlashAttribute("successMessage", "User created successfully!");
     return "redirect:/admin/user";
   }
 
@@ -105,14 +121,26 @@ public class UserController {
         : new User("None", "None",
             "None", "None",
             "None"));
+
+    // Không cần truyền xuống mỗi lân request nx do đã lưu trong session bằng
+    // @SessionAttribute
+    // model.addAttribute("roles", roleService.getAll());
+
     return "/admin/user/update";
   }
 
   @PostMapping("/admin/user/update")
-  public String updateUser(@ModelAttribute User user,
+  public String updateUser(@Valid @ModelAttribute User user, BindingResult userBindingResult,
       @RequestParam("avatar_file") MultipartFile file,
       @RequestParam(defaultValue = "false") boolean isDeleteAvatar,
       RedirectAttributes redirectAttributes) {
+
+    if (userBindingResult.hasFieldErrors()) {
+      List<FieldError> errors = userBindingResult.getFieldErrors();
+      errors.forEach((e) -> System.out.println(">>> ERR Update User: " + e.getField() + " - " + e.getDefaultMessage()));
+      return "admin/user/update";
+    }
+
     if (isDeleteAvatar) {
       fileService.deleteImage(user.getAvatar(), Type.AVATAR);
       user.setAvatar(null);
@@ -139,7 +167,7 @@ public class UserController {
   }
 
   @PostMapping("/admin/user/delete/{id}")
-  public String deleteUser(@PathVariable long id) {
+  public String deleteUser(@PathVariable long id, RedirectAttributes redirectAttributes) {
     User user = userService.get(id);
     if (user == null)
       return "redirect:/admin/user";
@@ -149,6 +177,7 @@ public class UserController {
       fileService.deleteImage(fileName, Type.AVATAR);
     }
     userService.delete(id);
+    redirectAttributes.addFlashAttribute("successMessage", "User deleted successfully!");
     return "redirect:/admin/user";
   }
